requirements.txt
langchain==0.0.277
python-dotenv==1.0.0
chromadb==0.3.29
sentence_transformers==2.2.2
pandas==2.0.3
openai==0.27.8
jq==1.5.0
